+++ Assets/Packages/HurricaneVR/Framework/Scripts/Components/HVRDamageProvider.cs
@@ -5,7 +5,7 @@ namespace HurricaneVR.Framework.Components
     public class HVRDamageProvider : MonoBehaviour
     {
         public WeaponType WeaponType;
-        public float Damage = 25f;
+        public int Damage = 25;
         public float Force;
 
         [Tooltip("Player transform for ai frameworks like emerald ai")]
